ai_service_base/ai_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.11.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.0
psycopg-pool>=3.2.0
